separate different routes, add controller passport.js, code refactoring
